Let users click verification email from anywhere
Not sure this is worth it, but usually you need to navigate to verification links in the browser that originated them so that you're logged in. But it's annoying when an email originates from an action on your phone, for example, and you click it from your laptop, and then it says you're not authorized.